pyproject.toml
[project]
name = "thoughtful_ai_support_agent"
version = "0.1.0"
description = "A customer support AI Agent for Thoughtful AI using Gradio, LangChain, and Asyncio."
authors = [
    {name = "\"Sreenivas Manyam Rajaram\"", email = "sreenivas_mr@yahoo.com"},
]

dependencies = [
    "gradio>=4.44.0",
    "langchain>=0.3.1",
    "openai>=1.49.0",
    "faiss-cpu>=1.8.0.post1",
    "numpy>=1.26.4",
    "scikit-learn>=1.5.2",
    "langchain-community>=0.3.1",
    "tiktoken>=0.7.0",
    "langchain-openai>=0.2.1",
    "setuptools>=75.1.0",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]

test = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "requests>=2.32.3",
    "selenium>=4.25.0",
    "playwright>=1.47.0",
    "pytest-asyncio>=0.24.0",
]

dev = [
    "ruff>=0.6.8",
    "pyright>=1.1.382.post0",
    "pdm-pep517>=1.1.4",
    "pre-commit>=3.8.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
start = "python src/ui.py"
lint = "ruff src tests"
type-check = "pyright src tests"
test = "pytest --cov=src tests/"
coverage = "pytest --cov=src tests/ --cov-report=html"

[project.urls]
Homepage = "https://github.com/yourusername/thoughtful_ai_support_agent"
